+++ CHANGELOG.md
@@ -9,17 +9,25 @@ All notable changes to this project will be documented in this file.
 - Add support for the `CarbonChat` plugin
 - Add custom ChatGuard prefix
 - Add color & format code system similar to emoji system
-- Add strike command
+- Add separate strike and remove-strike commands
+- Native support for Minecraft 1.21.4
+- Filtering for sensitive personal information
 
 ### Enhanced
 
 - Strike system and saving
+- Debugging system
+- Update dependencies
+
+### Fixed
+
+- Several minor bugs
 
 ## [1.0.1] - Latest Stable Release (11.10.2024)
 
 ### Added
 
-- First time messsage upon plugin install.
+- First time message upon plugin install.
 
 ### Fixed